heap: update fetch remote curio to use react-query
